requirements.txt
torch>=1.9.0
torchvision>=0.10.0
hydra-core>=1.1.0
omegaconf>=2.1.0
matplotlib>=3.4.0
numpy>=1.20.0
pillow>=8.0.0
tqdm>=4.60.0
einops>=0.3.0
tensorboard>=2.5.0
pytest>=6.0.0
requests>=2.31.0
